+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **[FIX]** `<Snackbar>`: Fix alignment on IE11.
 - **[NEW]** New prop `errorClassName` on `<TextField>` and `<Textarea>`
 - [...]
+- **[NEW]** Loading state for `ConfirmationModal` confirm button
 
 # v11.2.0 (04/09/2019)